Finding closest neighbor in column.
It is stated in the description that the edges matrix is symmetric and thus the function only uses the upper triangle. I found a bug where the code returns nothing if you are using a matrix where the lower triangle of the matrix is zero. This way, the function can be fed only the top triangle of the symmetric matrix as it checks both the row and column for its closest neighbor.